lang-agent-backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "lang-agent"
version = "0.1.0"
description = "Agents system using LangGraph"
authors = []
readme = "README.md"
repository = "https://github.com/cqzyys/lang-agent"
license = "MIT"

[tool.mypy]
disallow_untyped_defs = "True"

[tool.poetry.urls]
"Source Code" = ""
"Release Notes" = ""

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain-core = "^0.3.15"
langgraph = "^0.3.5"
langchain-openai = "^0.3.8"
langchain-milvus = "^0.1.8"
sqlalchemy = "^2.0.39"
py-xid = "^1.0.0"
grandalf = "^0.8"
langgraph-checkpoint-sqlite = "^2.0.6"
langchain-mcp-adapters = "^0.1.0"
nest-asyncio = "^1.6.0"
aiosqlite = "^0.21.0"
fastapi = "^0.115.14"
langchain-community = "^0.3.27"
pypdf = "^5.8.0"
nltk = "^3.9.1"
langchain-unstructured = "^0.1.6"
unstructured = {extras = ["docx", "md"], version = "^0.18.9"}
langchain-postgres = "^0.0.15"

[tool.ruff.lint]
select = ["E", "F", "I", "T201"]

[tool.coverage.run]
omit = ["tests/*"]

[tool.pytest.ini_options]
addopts = "--strict-markers --strict-config --durations=5"
markers = [
    "compile: mark placeholder test used to compile integration tests without running them",
]
asyncio_mode = "auto"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.codespell]
optional = true

[tool.poetry.group.lint]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-socket = "^0.7.0"
pytest-watcher = "^0.3.4"
langchain-tests = "^0.3.5"
pytest-cov = "^6.2.1"

[tool.poetry.group.codespell.dependencies]
codespell = "^2.2.6"

[tool.poetry.group.lint.dependencies]
ruff = "^0.5"

[tool.poetry.group.typing.dependencies]
mypy = "^1.10"
